Add redirect URI to MultipleFileInput component
A new attribute `redirect_uri` was added to the MultipleFileInput component to support redirection after file submission. This was implemented by introducing a new function `redirect_to_after_submit()` that allows setting a `redirect_uri`. On rendering the input field, this URI is set as an attribute 'data-redirect-uri'. This helps in enhancing user experience by redirecting them to an appropriate location after they have uploaded files.