Add duration input field and update dependent libraries
This commit adds a new duration input field in the form.rs page. This field aims to conveniently display durations using chrono's Duration data type. The commit also updates some dependent crates/packages in Cargo.toml and Cargo.lock files. The update includes adding the timeago dependency which is used to format the durations in a more readable format. Changes to field.rs simplify and improve the rendering of the Field struct.