Add 'stroked' option to IconPack and adjust experimental checks
Added a 'stroked' option to the IconPack in lib/build.rs and made changes to the viewy-icons.toml to demonstrate its usage. Additionally, adjusted the settings of RsExperimentalChecks in the Project_Default.xml file and commented out a block of code in form.js for testing. These modifications are performed to provide more customization options and to better handle form data.